ALMOST THERE

+++ QBot/Environments/States/State.py
@@ -13,7 +13,7 @@ class State:
         # Create array of zeros length max_objects by number of types of objects.
         encoded_arr = np.zeros((max_objects, num_classes), dtype=int)
         # Flip to 1's at all the indexes in index_list.
-        encoded_arr[np.arange(index_list.size), index_list] = 1
+        encoded_arr[np.arange(index_list.size, dtype=int), index_list] = 1
         return encoded_arr
         # return [[(1 if index_list[j] == i else 0) if j < len(index_list) else 0 for
         #   i in range(num_classes)] for j in range(max_objects)]
